Small fixes
- Localize the "Loading..." message when the page first loads (this isn't the final loading state, but this was 30 seconds to fix so figured might as well)
- Move the providing of `unfurlEndpoint` and `trustedSitesEndpoint` up one level, so that `d2l-activity-editor` doesn't need to know about it
