+++ components/d2l-activity-editor/lang/en.js
@@ -21,5 +21,6 @@ export default {
 	"setUngraded": "Set Ungraded", // Menu item for setting the activity to ungraded
 	"scoreOutOf": "Score Out Of", // ARIA label for the score out of field, when creating/editing an activity
 	"emptyScoreOutOfError": "A points value must be specified for activities in Grades", // Error message to inform user that the score out of value is a required field when a grade item is associated
-	"invalidScoreOutOfError": "Score Out Of must be greater than or equal to 0.01 and less than or equal to 9,999,999,999" // Error message when an invalid score out of value is entered
+	"invalidScoreOutOfError": "Score Out Of must be greater than or equal to 0.01 and less than or equal to 9,999,999,999", // Error message when an invalid score out of value is entered
+	"loading": "Loading...", // Message displayed while page is loading
 };
